backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1 
openai==1.76.0
